# docs/swagger.yaml
openapi: 3.0.3
info:
  title: hiauth 权限管理 API
  version: 1.0.0
  description: Hiauth permission management API.
servers:
  - url: http://localhost:8080
    description: Local
...
paths:
...
  /api/v1/system-configs:
    get:
      tags: [系统配置]
      summary: 查询系统配置列表
      parameters:
        - $ref: "#/components/parameters/Keyword"
        - $ref: "#/components/parameters/PageNo"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/SortField"
        - $ref: "#/components/parameters/SortOrder"
        - name: group
          in: query
          schema:
            type: string
          description: 配置分组
        - name: enabled
          in: query
          schema:
            type: integer
            enum: [0, 1]
          description: 是否启用，0 禁用，1 启用
      responses:
        "200":
          description: 系统配置分页列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemConfigPageResponse"
    post:
      tags: [系统配置]
      summary: 创建系统配置
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSystemConfigRequest"
      responses:
        "200":
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemConfigResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/system-configs/enabled:
    get:
      tags: [系统配置]
      summary: 查询启用的系统配置列表
      parameters:
        - name: group
          in: query
          schema:
            type: string
          description: 配置分组，不传返回所有启用配置
      responses:
        "200":
          description: 启用配置列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemConfigListResponse"
  /api/v1/system-configs/enabled-map:
    get:
      tags: [系统配置]
      summary: 查询启用的系统配置 Map
      parameters:
        - name: group
          in: query
          schema:
            type: string
          description: 配置分组，不传返回所有启用配置
      responses:
        "200":
          description: 启用配置键值对
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemConfigMapResponse"
  /api/v1/system-configs/by-key/{key}:
    get:
      tags: [系统配置]
      summary: 根据配置键查询系统配置
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
          description: 配置键
      responses:
        "200":
          description: 系统配置详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemConfigResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/system-configs/{id}:
    get:
      tags: [系统配置]
      summary: 查询系统配置详情
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: 系统配置详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemConfigResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
    put:
      tags: [系统配置]
      summary: 更新系统配置
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSystemConfigRequest"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemConfigResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
    delete:
      tags: [系统配置]
      summary: 删除系统配置
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/audit-logs:
    get:
      tags: [操作审计]
      summary: 查询操作审计列表
      description: 查询操作审计主记录，详情中的数据库变更前后数据请调用审计详情接口查看。
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Keyword"
        - $ref: "#/components/parameters/PageNo"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/SortField"
        - $ref: "#/components/parameters/SortOrder"
        - name: operatorName
          in: query
          schema:
            type: string
          description: 操作人用户名，支持模糊查询
        - name: module
          in: query
          schema:
            type: string
            example: 用户管理
          description: 操作模块
        - name: action
          in: query
          schema:
            type: string
            enum: [create, update, delete, authorize, query, login, logout]
          description: 操作类型，create 新增，update 修改，delete 删除，authorize 授权，query 查询，login 登录，logout 退出登录
        - name: resourceType
          in: query
          schema:
            type: string
            example: user
          description: 资源类型，例如 user、role、department、menu、system_config、user_roles、role_menus
        - name: resourceId
          in: query
          schema:
            type: integer
            format: int64
          description: 资源 ID
        - name: status
          in: query
          schema:
            type: string
            enum: [success, fail]
          description: 审计状态，当前数据变更审计主要记录 success
        - name: startTime
          in: query
          schema:
            type: string
            example: "2026-05-22 00:00:00"
          description: 操作开始时间，格式 yyyy-MM-dd HH:mm:ss，按 createdAt 过滤
        - name: endTime
          in: query
          schema:
            type: string
            example: "2026-05-22 23:59:59"
          description: 操作结束时间，格式 yyyy-MM-dd HH:mm:ss，按 createdAt 过滤
      responses:
        "200":
          description: 操作审计分页列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogPageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/audit-logs/{id}:
    get:
      tags: [操作审计]
      summary: 查询操作审计详情
      description: 返回操作审计主记录及数据库变更明细，包括 beforeData、afterData 和 changedFields。
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: 操作审计详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: 主键 ID
    Keyword:
      name: keyword
      in: query
      schema:
        type: string
      description: 关键字
    PageNo:
      name: pageNo
      in: query
      schema:
        type: integer
        default: 1
      description: 页码
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        default: 10
        maximum: 200
      description: 每页条数
    SortField:
      name: sortField
      in: query
      schema:
        type: string
      description: 排序字段
    SortOrder:
      name: sortOrder
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: desc
      description: 排序方向
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Response"
  schemas:
    Response:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: OK
        data:
          nullable: true
      required: [code, message]
...
    BaseModel:
      type: object
      properties:
        id:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
          nullable: true
        createdBy:
          type: string
        updatedAt:
          type: string
          format: date-time
          nullable: true
        updatedBy:
          type: string
...
    SystemConfig:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - type: object
          properties:
            key:
              type: string
              example: site.title
            value:
              type: string
              example: hiauth
            name:
              type: string
              example: 系统名称
            valueType:
              type: string
              enum: [string, number, bool, json]
            group:
              type: string
              example: basic
            description:
              type: string
            enabled:
              type: integer
              enum: [0, 1]
            sort:
              type: integer
    AuditChange:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - type: object
          properties:
            auditLogId:
              type: integer
              format: int64
              description: 关联的审计日志 ID
            tableName:
              type: string
              example: user
              description: 发生变更的数据表
            recordId:
              type: integer
              format: int64
              description: 发生变更的数据主键 ID
            operation:
              type: string
              enum: [create, update, delete]
              description: 数据变更类型
            beforeData:
              type: string
              description: 变更前数据 JSON 字符串，新增时为空；敏感字段会脱敏
              example: "{\"id\":1,\"username\":\"admin\"}"
            afterData:
              type: string
              description: 变更后数据 JSON 字符串，删除时为空；敏感字段会脱敏
              example: "{\"id\":1,\"username\":\"admin\",\"nickname\":\"管理员\"}"
            changedFields:
              type: string
              description: 发生变化的字段名 JSON 数组字符串
              example: "[\"nickname\"]"
            fieldChanges:
              type: string
              description: 字段变更明细 JSON 数组字符串，仅包含业务字段；每项包含 field、before、after
              example: "[{\"field\":\"nickname\",\"before\":\"张三\",\"after\":\"李四\"}]"
    AuditLog:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - type: object
          properties:
            requestId:
              type: string
              description: 请求链路 ID
            operatorId:
              type: integer
              format: int64
              description: 操作人用户 ID
            operatorName:
              type: string
              description: 操作人用户名
            module:
              type: string
              example: 用户管理
              description: 操作模块
            action:
              type: string
              enum: [create, update, delete, authorize, query, login, logout]
              description: 操作类型
            resourceType:
              type: string
              example: user
              description: 资源类型
            resourceId:
              type: integer
              format: int64
              description: 资源 ID
            description:
              type: string
              example: 修改用户，ID：1
              description: 操作描述
            method:
              type: string
              example: PUT
              description: HTTP 方法
            path:
              type: string
              example: /api/v1/users/1
              description: 请求路径
            ip:
              type: string
              description: 客户端 IP
            userAgent:
              type: string
              description: User-Agent
            status:
              type: string
              enum: [success, fail]
              description: 操作状态
            errorMessage:
              type: string
              description: 失败原因
            durationMs:
              type: integer
              format: int64
              description: 请求耗时，单位毫秒
            changes:
              type: array
              description: 数据库变更明细，列表接口通常为空，详情接口返回
              items:
                $ref: "#/components/schemas/AuditChange"
...
    PageSystemConfig:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: "#/components/schemas/SystemConfig"
        total:
          type: integer
          format: int64
        pageNo:
          type: integer
        pageSize:
          type: integer
    PageAuditLog:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"
        total:
          type: integer
          format: int64
        pageNo:
          type: integer
        pageSize:
          type: integer
...
    CreateSystemConfigRequest:
      type: object
      required: [key, name]
      properties:
        key:
          type: string
          description: 配置键，支持字母、数字、下划线、点、冒号、短横线，必须以字母开头
          example: site.title
        value:
          type: string
        name:
          type: string
        valueType:
          type: string
          enum: [string, number, bool, json]
          default: string
        group:
          type: string
          default: default
        description:
          type: string
        enabled:
          type: integer
          enum: [0, 1]
          default: 1
        sort:
          type: integer
          default: 0
        operator:
          type: string
    UpdateSystemConfigRequest:
      allOf:
        - $ref: "#/components/schemas/CreateSystemConfigRequest"
...
    SystemConfigResponse:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/SystemConfig"
    SystemConfigPageResponse:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/PageSystemConfig"
    AuditLogResponse:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/AuditLog"
    AuditLogPageResponse:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/PageAuditLog"
    SystemConfigListResponse:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/SystemConfig"
    SystemConfigMapResponse:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data:
              type: object
              additionalProperties:
                type: string
...
    MessageResponse:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data:
              nullable: true